setup a basic URL interceptor that can muck with requests
Right now we're just adding a DNT header and forcing all cookies to
session expiration, though this will change later.

It took a ton of time to figure out why I wasn't seeing
http->https redirections in -connection:willSendRequest:redirectResponse:
and only saw them when the hostname changed in a redirection.  This
made it difficult to track whether the current URL was secure or not
for updating the search bar icon.

A workaround for this is storing the original request as we get it
startLoading and then comparing it to the request we get later to
figure out if it's a different request (since apparenly this
redirection happens behind the scenes).