add a license
Based on a 3-clause BSD license allowing use and modification in
source forms, and use in binary form, but explicitly forbidding
redistribution in binary form.

This should not negatively affect anyone's personal use of this code
unless you're trying to rip me off by changing the name and icon and
sticking it on the App Store for $1.99.