add a "tabs" implementation
Put each UIWebView on a main horizonally-scrollable UIScrollView
next to each other.  When hitting the tabs button, zoom out on each
one and allow paging through them.

Bind URL requests to each WebViewTab.

Still needs work to make a "context" toolbar when zoomed out to add
a "+" button, and to detect javascript window.open and <a
target="_blank"> to open in a new tab.