give in and (ab)use the user-agent to store the tab number
From my internet diggings, the only way to modify an NSURLRequest in
a UIWebView that does not come through its delegate
shouldStartLoadWithRequest: method (as in the case of page element
requests like images, fonts, etc.) is to set the user agent before
creating a UIWebView.  Append the WebViewTab hash onto the
User-Agent header before creating it, and then in URLInterceptor,
look up the hash to find the corresponding WebViewTab and strip off
the hash from the UA before sending it to the server.

While this does leave the hash in navigator.userAgent, it shouldn't
really affect anything as far as browser identification or leaking
anything sensitive.

Give up on the other methods of correlating a request in
URLInterceptor to a WebViewTab and make non-matching fatal again.