add an HTTP Strict Transport Security (RFC6797) implementation
It wasn't until I was about 90% finished with all of this that I
realized that WebKit already has its own HSTS implementation behind
the scenes that caches entries and has a preload list of a few
hundred domains.  Oh well.

Since we can't manipulate that list, continue with our own and
preload it with Chromium's preload list (which Mozilla also appears
to use), which currently has 1427 hosts and is freely licensed.