restore tabs more efficiently
Don't refresh every tab when recreating it, just mark that it needs
to be refreshed.  Only refresh the current tab and when another
restored tab becomes current, it will get refreshed automatically.

Start storing the title with the restoration state so when viewing
the tab list, one can see what a tab is for before it gets
refreshed.