bring back EV org name display and add an SSL cert viewer
CKHTTPConnection calls out to URLInterceptor's delegate method
HTTPConnection:shouldContinueWithSecTrustRef: to find out if it
should allow connection to the SSL host (this always returns true
for now, but could be used to blacklist certs later).

In that call, WebViewTab caches an SSLCertificate object which is
built from the actual certificate data received from the host.
SSLCertificate uses DTFoundation's ASN1 parser to extract some
useful bits about the cert and save it for later.

When tapping on the lock icon now, it pops up a simple list of
attributes about the cert, highlighting EV org names and weak (sha1)
certificates.

Note that the ASN1/X509 parsing done here is only to extract text
information to be displayed to the end user and is not used in any
security-related context like deciding whether to trust a
certificate.  All of that is still handled automatically by
UIWebView's guts.