WVC: set our user agent to be more like safari
That is, actually contain the word "Safari", insert a fake safari
version number based on our ios version, and append "Safari/" with
the webkit build number.

Should make Endless more complaint with certain websites that check
for the UA to contain "Safari", and makes us blend in with other
traffic.