CKHTTPConnection: fix problem with keep-alive connections
Sometimes CFReadStreamCreateForHTTPRequest with
kCFStreamPropertyHTTPAttemptPersistentConnection set will give us a
socket that was used on a previous connection to this host but is
dead, so when we try to do anything on it, we just get back
"connection reset by peer" right away and the request fails.

If we get a stream event of NSStreamEventErrorOccurred without ever
seeing that the stream was readable or writable, close the socket
and retry it once before sending the error back upstream.

This fixes the long-standing bug with loading m.imgur.com pages
where it would load everything, pause half a second and then refresh
itself, removing the title bar from the image.  This was caused by
it doing two XHR requests to api.imgur.com when the page loaded, and
this connection bug would cause the second one to fail right away.
The imgur JS code would retry this request after the page loaded and
then succeed, causing the redraw (which seems like a bug in their JS
code, but whatever).

This was tracked down with the new DummyURLInterceptor code.