Fix setting minimum TLS version to actually support ranges instead of one value
Using kCFStreamSSLLevel sets the required version instead of the
minimum supported version, so when downgrading to TLS 1.0, it forces
everything to 1.0 instead of supporting 1.1 and 1.2 if the server
supports it.

For TLS 1.2, keep using kCFStreamSSLLevel, but for TLS 1.0/1.1/1.2,
or SSL v3, use SSLSetProtocolVersionMin/Max to create ranges of
support.  In all cases, keep TLS 1.2 as the highest supported
version so if the server supports it, it should keep using that.

Also, for some reason that TLS cache thing is no longer a problem,
so remove the popup warning after changing the per-host value.