URLInterceptor: do smarter CSP header manipulation
for a site like mobile.twitter.com that already has a CSP header
with a frame-src specification, we won't be able to open our
endlessipc:// IPC urls to override window.open because when our
injected javascript tries to do the IPC call in an iframe, webkit
enforces the CSP and the new window open fails (and a CSP report is
generated).

add a parser of CSP headers that can break up each directive and
modify it to add or prepend endlessipc: for frame-src and child-src.