WVC: add a wrapper view between root view and scroller
In dark mode, the tab toolbar was not completely transparent and had
a slightly different color, despite setting its backgroundColor and
barTintColor to the same shade of gray as the tab chooser.

To get around this, add blank images as the background which makes
it completely transparent.  But then when in light mode with the
toolbar at the bottom, the window is black, which made the toolbar
black (the tab chooser can't extend to the full height of the
window).

Work around /that/ by adding a wrapper view between the root view
and the scroller, that will remain the same size as the root view
minus its status bar.  This lets the root view become black but show
a light gray background behind the tab toolbar.