CKHTTPConnection: finally fix keep-alive issue and re-enable
We can only disable weak ciphers if we're in kSSLIdle, which we
won't be in if we're re-using a persistent connection that has
already negotiated ciphers.

Now that we can use persistent connections reliably, remove the
conditional define.

Also, while we're here, remove the block setting TLS 1.0 and 1.1.
iOS 10 doesn't do SSL 2/3 anymore, so there's no need to specify any
settings unless the host has TLS 1.2 required.

closes #26