Support raw hardware keyboard input on the iPad
This is something WKWebView gets for free, but since we're still on
UIWebView, all of this work is required.

Bind all possible key permutations via UIKeyCommand, but depending
on the current first responder (i.e., whether it's a text field or
not), only return the global key shortcuts or our big list of all
keys.

If a key doesn't match the list of global shortcuts, pass it to the
current WebViewTab where it is mapped to a Javascript
keycode/charcode and injected as a KeyboardEvent, overriding some
properties that initKeyboardEvent doesn't seem to setup which some
websites/JS libraries need.

Scrolling is achieved by checking the result of the keydown event
and if it wasn't canceled by the site's Javascript, do smooth window
scrolling in our injected JS (using requestAnimationFrame to be CPU
friendly).

Tested on GitHub (scrolling with j, k), den.im (n, m), JSLinux (full
keyboard input including control+a, etc.), and a few other
javascript terminal emulator things.  Also tested with Apple's
bluetooth keyboard which adds the Escape and F1-F12 keys.