URLInterceptor: fix CSP header munging with script-src nonce
When a site has an existing CSP with script-src but has things in it
like 'unsafe-inline', when we prepend our nonced version of the
injected script, the browser effectively ignores all other values in
the directive and requires that all of them have nonces.  Since none
of them do, all of the 'unsafe-inline' scripts on the page break.

Fixes images loading on a medium.com page which has such a CSP.