WebViewTab: implement forced refresh better
NSURLRequestReloadIgnoringLocalCacheData doesn't really do much
since it still sends the request to the server with an If-None-Match
header, causing a 304 response instead of 200.

Passing other types of cachePolicy values doesn't seem to have any
effect, so hack around this by setting a flag on the WebViewTab for
the duration of the reload, and cause URLInterceptor to delete any
If-None-Match headers that were automatically added when this flag
is set.