WVT: add Universal Link protection
Universal Links in iOS allow 3rd party apps to claim URL hosts that
they own, so when links to those hosts are being opened in other
apps, that 3rd party app is executed to handle that URL request.

Unfortunately there is no way to disable this, so in a web browser
app like Endless (I also tested in Chrome, Firefox, Brave, Tob, and
Onion Browser), tapping on a link can immediately spawn an installed
3rd party app which will make that URL request without any
confirmation or warning.

For example, if the user has the eBay app installed and taps on a
link in Endless pointing to a http://rover.ebay.com/ URL, the eBay
app will immediately be opened and show the auction page being
requested, which could contain an <img> tag in the auction
description that loads from a 3rd party server.  While this isn't a
big deal for Endless, it is for Tor- and VPN-based apps that are
based on Endless which are trying to keep the user's network
activity contained inside the app.

Since UIWebView (and WKWebView) offer no indication that such a URL
will be opened as a Universal Link (and probably won't ever, for the
same reason that iOS disabled UIApplication:canOpenURL: so apps
can't figure out which other apps the user has installed), implement
a workaround.

In WebViewTab's webView:shouldStartLoadWithRequest: delegate method,
always return NO for top-level requests that get here (which are
links that have been tapped on, window.location= calls, and iframes)
but then just start a new request for the same URL.  This seems
enough to bypass Universal Link activation and still works with a
bunch of sites and Javascript that I tested.

Test URL: https://endl.es/tests/decloak