remove FakeWindow and FakeLocation getters
Once a tab opens a window, it should only be able to set that tab's
location and name, and not get any attributes from it.  This matches
Firefox's behavior.

A same-origin window should remain accessible, but since there is no
convenient, secure way to detect whether the URL should be
considered by UIWebView as same-origin, just leave these
unimplemented for now.