rework and reorganize `OnionManager` a bit.
now we've got `OnionManagerDelegate` which is registered as a callback when
`OnionManager startTor` happens. callback currently passes back bootstrap
progress and the "finished" (progress==100) event.

`OBRootViewController` can't be rewritten in swift because of the bad hack
in `-userFinishedConnecting` where we go back and dynamically push to
Endless' original `webViewController` (which was the original root view
controller for Endless). i think? maybe i am doing the cocoapods or
headers wrong.